Refactor __init__.py
